+++ src/CMakeLists.txt
@@ -5,7 +5,7 @@ ornstein_uhlenbeck.cpp
 core.cpp
 gaussian.cpp
 gsl_errors.cpp
-filter.cpp
+kca.cpp
 states.cpp
 states_exceptions.cpp
 linalg.cpp
@@ -26,7 +26,10 @@ trading_levels.cpp
 trading_levels_exponential.cpp
 trading_levels_params.cpp
 type_conversion.cpp
-interface.cpp)
+entrypoint_optimal_trading_levels.cpp
+entrypoint_kca_filter.cpp
+entrypoint_ou_model.cpp
+)
 
 target_include_directories(stochastic_models
 PRIVATE
